Rename react-redux-boilerplate dir to src
