+++ src/components/QuoteSection.jsx
