+++ src/components/QuoteSection.less
