+++ src/reducers/quote.js
