+++ src/test/components/CounterSection_test.jsx
