+++ src/test/components/QuoteSection_test.jsx
